Refactor methods/structs better
